Fix composer version detection
This is in case a warning is shown before the actual composer version. In the case where XDebug is installed the following error is shown :
```
Could not find composer version from version string: You are running composer with xdebug enabled. This has a major impact on runtime performance. See https://getcomposer.org/xdebug
Composer version 1.2.0 2016-07-19 01:28:52
```