Add support for using the system tar binary
This adds an option (-s/--systemTar) which makes the archiving and
extracting use the system tar binary instead of tar-fs. This of course
requires that the system has tar installed, which is why it is added as
an option and not used by default.

In my experience, this increases the performance of the extraction by
around three to six times.