Fix npmConfig for lock versions 2 and 3
The `getFileHash` function has been created or only tested against
package-lock.json files with a lock version at 1. The version 2
introduced the `packages` object, which contains the project's
metadata, including `dependencies` and `devDependencies`.

This commit fix the hash generation that was previously done on
an empty string. The consequence was no cache invalidation since
the hash was always the same in case of a lock version 2 or 3.

The hash generation can easily be tested by generating different
package-lock.json file versions with the folowing command in an
npm project: `npm i --lockfile-version 1 --package-lock-only`.